Move some account types to user
