+++ src/api/index.ts
@@ -1,4 +1,3 @@
-export * from './accounts';
 export * from './api';
 export * from './biblib';
 export * from './export';
@@ -8,8 +7,10 @@ export * from './models';
 export * from './reference';
 export * from './search';
 export * from './types';
+export * from './user';
 export * from './vault';
 export * from './vis';
 
 import api from './api';
+
 export default api;
